test(apm): strengthen tracer assertions and resolver ownership
