+++ packages/core/src/tracing/datadog.test.ts
@@ -37,10 +37,10 @@ describe('resolveDatadogTracer', () => {
 describe('best-effort tracing', () => {
   it('forwards explicit resource and measurement without changing the operation or tags', () => {
     const tags = { 'span.kind': 'server' };
+    const calls: { name: string; options: Parameters<DatadogTracer['trace']>[1] }[] = [];
     const tracer: DatadogTracer = {
       trace(name, options, work) {
-        expect(name).toBe('mcp.tool');
-        expect(options).toEqual({ resource: 'agor_test', measured: true, tags });
+        calls.push({ name, options });
         return work();
       },
     };
@@ -50,6 +50,12 @@ describe('best-effort tracing', () => {
         measured: true,
       })
     ).toBe(42);
+    expect(calls).toEqual([
+      {
+        name: 'mcp.tool',
+        options: { resource: 'agor_test', measured: true, tags },
+      },
+    ]);
   });
 
   it('runs work once when disabled or tracing throws before/after invocation', async () => {
